requirements.txt
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
requests>=2.31.0